DEV-2420: Move logic in init into sub-modules.
