+++ src/gdcmodels/esmodels/case_centric/mapping.yaml
