+++ src/gdcmodels/esmodels/case_set/mapping.yaml
