+++ src/gdcmodels/esmodels/file_set/mapping.yaml
