+++ src/gdcmodels/esmodels/file_set/vestigial.yaml
