+++ src/gdcmodels/esmodels/gdc_from_graph/case/mapping.yaml
