+++ src/gdcmodels/esmodels/gdc_from_graph/file/vestigial.yaml
