+++ src/gdcmodels/esmodels/ssm_centric/mapping.yaml
