+++ src/gdcmodels/esmodels/ssm_centric/vestigial.yaml
