+++ src/gdcmodels/esmodels/ssm_occurrence_centric/mapping.yaml
