+++ src/gdcmodels/esmodels/ssm_occurrence_centric/vestigial.yaml
