+++ src/gdcmodels/esmodels/ssm_set/vestigial.yaml
